mdadm/grow: correct size and chunk_size casting
With commit 4b74a905a67e
("mdadm/grow: Component size must be larger than chunk size") mdadm returns
incorrect message if size given to grow was greater than 2 147 483 647 K.
Cast chunk_size to "unsigned long long" instead of casting size to "int".

+++ Grow.c
@@ -1821,7 +1821,7 @@ int Grow_reshape(char *devname, int fd,
 	}
 
 	if (array.level > 1 && s->size > 1 &&
-	    (array.chunk_size / 1024) > (int)s->size) {
+	    (unsigned long long) (array.chunk_size / 1024) > s->size) {
 		pr_err("component size must be larger than chunk size.\n");
 		return 1;
 	}
